requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-Login==0.6.3
python-dotenv==1.0.0
requests==2.31.0
pyTelegramBotAPI==4.16.1
SQLAlchemy==2.0.23
WTForms==3.1.1
email-validator==2.1.0